+++ packages/select/src/components/ui/MultiSelect.stories.tsx
@@ -1,15 +1,15 @@
 import { Box, Indent, Spacing } from "@stenajs-webui/core";
 import * as React from "react";
-import { Select } from "./Select";
+import { MultiSelect } from "./MultiSelect";
 
 export default {
-  title: "select/Select",
-  component: Select,
+  title: "select/MultiSelect",
+  component: MultiSelect,
 };
 
 export const Standard = () => (
   <div style={{ width: "400px" }}>
-    <Select
+    <MultiSelect
       options={[
         {
           value: "Mattias",
@@ -28,61 +28,11 @@ export const Standard = () => (
   </div>
 );
 
-export const LongLabels = () => (
-  <div style={{ width: "400px" }}>
-    <Select
-      options={[
-        {
-          value: "lorem",
-          label:
-            "Lorem ipsum dolor sit amet, consectetur adipiscing elit. Praesent purus massa, molestie ac sapien in, sodales interdum arcu",
-        },
-        {
-          value: "Johan",
-          label: "Johan",
-        },
-        {
-          value: "Dennis the menace",
-          label: "Dennis the menace",
-        },
-      ]}
-    />
-  </div>
-);
-
-export const StandardDark = () => (
-  <div style={{ width: "400px" }}>
-    <Box background={"#2e4662"}>
-      <Indent num={4}>
-        <Spacing num={4}>
-          <Select
-            options={[
-              {
-                value: "Mattias",
-                label: "Mattias",
-              },
-              {
-                value: "Johan",
-                label: "Johan",
-              },
-              {
-                value: "Dennis the menace",
-                label: "Dennis the menace",
-              },
-            ]}
-            variant={"dark"}
-          />
-        </Spacing>
-      </Indent>
-    </Box>
-  </div>
-);
-
 export const WithGroupHeadings = () => (
   <div style={{ width: "400px" }}>
     <Indent num={4}>
       <Spacing num={4}>
-        <Select
+        <MultiSelect
           options={[
             {
               label: "Group heading 1",
@@ -117,9 +67,37 @@ export const WithGroupHeadings = () => (
   </div>
 );
 
+export const MultiselectDark = () => (
+  <div style={{ width: "600px" }}>
+    <Box background={"#2e4662"}>
+      <Indent num={4}>
+        <Spacing num={4}>
+          <MultiSelect
+            options={[
+              {
+                value: "Mattias",
+                label: "Mattias",
+              },
+              {
+                value: "Johan",
+                label: "Johan",
+              },
+              {
+                value: "Dennis the menace",
+                label: "Dennis the menace",
+              },
+            ]}
+            variant={"dark"}
+          />
+        </Spacing>
+      </Indent>
+    </Box>
+  </div>
+);
+
 export const Disabled = () => (
   <div style={{ width: "400px" }}>
-    <Select
+    <MultiSelect
       options={[
         {
           value: "Mattias",