+++ packages/select/src/components/ui/Select.tsx
@@ -4,7 +4,7 @@ import SelectComponent, { mergeStyles, Props } from "react-select";
 import { defaultSelectTheme, selectThemeDark } from "../../SelectTheme";
 import { createStylesFromTheme } from "../../util/StylesBuilder";
 
-export interface SelectProps<T> extends Props<T> {
+export interface SelectProps<T> extends Props<T, false> {
   variant?: "dark" | "light";
 }
 
@@ -14,7 +14,7 @@ export const Select = <T extends {}>({
   ...selectProps
 }: SelectProps<T>) => {
   const selectStyles = useMemo(() => {
-    const sourceStyles = createStylesFromTheme(
+    const sourceStyles = createStylesFromTheme<T, false>(
       variant === "light" ? defaultSelectTheme : selectThemeDark
     );